refactoring sling
