+++ pkg/httpapi/models.go
@@ -44,3 +44,48 @@ type InspectionReportExportCompletionResponse struct {
 	Status string `json:"status"`
 	URL    string `json:"url,omitempty"`
 }
+
+type activityResponse struct {
+	Type     string            `json:"type"`
+	Metadata map[string]string `json:"metadata"`
+}
+
+// GetAccountsActivityLogResponse is the response from activity log history api
+type GetAccountsActivityLogResponse struct {
+	Activities    []activityResponse
+	NextPageToken string `json:"next_page_token"`
+}
+
+// GetAccountsActivityLogRequestParams contains fields required to make a post request to activity log history api
+type GetAccountsActivityLogRequestParams struct {
+	OrgID     string                    `json:"org_id"`
+	PageSize  int                       `json:"page_size"`
+	PageToken string                    `json:"page_token"`
+	Filters   accountsActivityLogFilter `json:"filters"`
+}
+
+// accountsActivityLogFilter filter for AccountsActivityLog
+type accountsActivityLogFilter struct {
+	Timeframe  timeFrame `json:"timeframe,omitempty"`
+	EventTypes []string  `json:"event_types"`
+	Limit      int       `json:"limit"`
+}
+
+type timeFrame struct {
+	From time.Time `json:"from"`
+}
+
+// NewGetAccountsActivityLogRequest build a request for AccountsActivityLog
+// for now it serves the purposes only for inspection.deleted. If we need later, we can change this builder
+func NewGetAccountsActivityLogRequest(pageSize int, from time.Time) *GetAccountsActivityLogRequestParams {
+	return &GetAccountsActivityLogRequestParams{
+		PageSize: pageSize,
+		Filters: accountsActivityLogFilter{
+			Timeframe: timeFrame{
+				From: from,
+			},
+			Limit:      pageSize,
+			EventTypes: []string{"inspection.deleted"},
+		},
+	}
+}
